Add spider.middleware/with-middleware
Following a discussion around improving the state of the web world, it
was mentioned that adding a more sane way of handling web middleware
would be useful. Currently we thread middleware with ->>, which results
in handlers executing in reverse order, which is incredibly confusing.
I think a preferable solution is listing a sequence of middleware to
execute in order, along with the final handler function. The new
higher-order `spider.middleware/with-middleware` function should help in
constructing these functions.