Clean up code, change nil handling in read-{edn,json}-body
Most notably, this commit changes the behavior of read-edn-body and
read-json-body to properly handle nil/empty values and return nil in
turn. The pattern of

```clj
(some-> request :body slurp attempt-read-edn)
```

is very common, and now cleanly handled in Spider. The Clojure version
is bumped up to 1.7.0 which is a requirement for framed.std.

Code is also cleaned up to destructure common request fields, clean up
whitespace, etc.